requirements.txt
pyyaml
pytest

# RL / training (Phase 3)
gymnasium
stable-baselines3
torch
numpy
